+++ CHANGELOG.rst
@@ -13,11 +13,15 @@ Unreleased
 **Bug fix:**
 
 - The :meth:`score` method of :class:`~glum.GeneralizedLinearRegressor` and :class:`~glum.GeneralizedLinearRegressorCV` now accepts offsets.
+- Fixed the calculation of the information matrix for the Binomial distribution with logit link, which affected nonrobust standard errors.
 
 **Other:**
 
 - The CI now runs daily unit tests against the nightly builds of numpy, pandas and scikit-learn.
 - ``oldest-supported-numpy`` is used for build. 
+- :meth:`~glum.GeneralizedLinearRegressor.std_errors` and :meth:`~glum.GeneralizedLinearRegressor.covariance_matrix`
+  of :class:`~glum.GeneralizedLinearRegressor` now accept data frames with categorical data or high sparsity.
+- Updated ``environment.yml`` and ``conda.recipe/meta.yaml`` to use tabmat 3.1.0.
 
 **New feature:**